Integrate the WebstackClientInfo into the WebstackClient
